Move timetable from MainActivity to TimetableFragment
